+++ voicevox_engine/utility/__init__.py
